refactored SIAlertView "god class" into separate .h/.m files.
