+++ SIAlertView/SIAlertItem.h
@@ -1,0 +1,28 @@
+//
+//  SIAlertItem.h
+//  SIAlertView
+//
+//  Created by Kevin Cao on 13-4-29.
+//  Copyright (c) 2013年 Sumi Interactive. All rights reserved.
+//
+
+#import <Foundation/Foundation.h>
+
+@class SIAlertView;
+
+typedef NS_ENUM(NSInteger, SIAlertViewButtonType) {
+    SIAlertViewButtonTypeDefault = 0,
+    SIAlertViewButtonTypeDestructive,
+    SIAlertViewButtonTypeCancel
+};
+
+typedef void(^SIAlertViewHandler)(SIAlertView *alertView);
+
+
+@interface SIAlertItem : NSObject
+
+@property (nonatomic, copy) NSString *title;
+@property (nonatomic, assign) SIAlertViewButtonType type;
+@property (nonatomic, copy) SIAlertViewHandler action;
+
+@end
