+++ SIAlertView/SIAlertItem.m
@@ -1,0 +1,13 @@
+//
+//  SIAlertItem.m
+//  SIAlertView
+//
+//  Created by Kevin Cao on 13-4-29.
+//  Copyright (c) 2013年 Sumi Interactive. All rights reserved.
+//
+
+#import "SIAlertItem.h"
+
+@implementation SIAlertItem
+
+@end
