+++ SIAlertView/SIAlertView.h
@@ -7,22 +7,14 @@
 //
 
 #import <UIKit/UIKit.h>
+#import "SIAlertBackgroundWindow.h"
+#import "SIAlertItem.h"
 
-extern NSString *const SIAlertViewWillShowNotification;
-extern NSString *const SIAlertViewDidShowNotification;
-extern NSString *const SIAlertViewWillDismissNotification;
-extern NSString *const SIAlertViewDidDismissNotification;
+extern NSString * const SIAlertViewWillShowNotification;
+extern NSString * const SIAlertViewDidShowNotification;
+extern NSString * const SIAlertViewWillDismissNotification;
+extern NSString * const SIAlertViewDidDismissNotification;
 
-typedef NS_ENUM(NSInteger, SIAlertViewButtonType) {
-    SIAlertViewButtonTypeDefault = 0,
-    SIAlertViewButtonTypeDestructive,
-    SIAlertViewButtonTypeCancel
-};
-
-typedef NS_ENUM(NSInteger, SIAlertViewBackgroundStyle) {
-    SIAlertViewBackgroundStyleGradient = 0,
-    SIAlertViewBackgroundStyleSolid,
-};
 
 typedef NS_ENUM(NSInteger, SIAlertViewTransitionStyle) {
     SIAlertViewTransitionStyleSlideFromBottom = 0,
@@ -32,8 +24,6 @@ typedef NS_ENUM(NSInteger, SIAlertViewTransitionStyle) {
     SIAlertViewTransitionStyleDropDown
 };
 
-@class SIAlertView;
-typedef void(^SIAlertViewHandler)(SIAlertView *alertView);
 
 @interface SIAlertView : UIView
 
@@ -57,6 +47,10 @@ typedef void(^SIAlertViewHandler)(SIAlertView *alertView);
 @property (nonatomic, strong) UIFont *buttonFont NS_AVAILABLE_IOS(5_0) UI_APPEARANCE_SELECTOR;
 @property (nonatomic, assign) CGFloat cornerRadius NS_AVAILABLE_IOS(5_0) UI_APPEARANCE_SELECTOR; // default is 2.0
 @property (nonatomic, assign) CGFloat shadowRadius NS_AVAILABLE_IOS(5_0) UI_APPEARANCE_SELECTOR; // default is 8.0
+
+- (void)setup;
+- (void)invaliadateLayout;
+- (void)resetTransition;
 
 - (id)initWithTitle:(NSString *)title andMessage:(NSString *)message;
 - (void)addButtonWithTitle:(NSString *)title type:(SIAlertViewButtonType)type handler:(SIAlertViewHandler)handler;
