+++ SIAlertView/SIAlertViewController.h
@@ -1,0 +1,18 @@
+//
+//  SIAlertViewController.h
+//  SIAlertView
+//
+//  Created by Kevin Cao on 13-4-29.
+//  Copyright (c) 2013年 Sumi Interactive. All rights reserved.
+//
+
+#import <UIKit/UIKit.h>
+
+@class SIAlertView;
+
+@interface SIAlertViewController : UIViewController
+
+@property (nonatomic, strong) SIAlertView *alertView;
+
+@end
+
